chore: enable strict mypy and add CI workflow
- Add [tool.mypy] strict=true to pyproject.toml with ddgs override
- Target all bare Agent / RunContext / RunContext[Any] annotations
- Drop @runtime_checkable from Protocol classes (incompatible with
  isinstance checks that mypy flags under strict mode)
- Remove stale # type: ignore comments now that strict mypy is enforced
- Add explicit str() / float() / cast() where needed for type narrowing
- Add .github/workflows/ci.yml: ruff + mypy lint, pytest matrix
  (3.10–3.12)

+++ lovia/reliability.py
@@ -160,4 +160,4 @@ class RetryPolicy:
     def backoff_delay(self, attempt: int) -> float:
         """Jittered exponential delay before retrying after ``attempt`` failures."""
         delay = min(self.backoff_max, self.backoff_base * (2 ** (attempt - 1)))
-        return delay * (0.5 + random.random())
+        return float(delay * (0.5 + random.random()))
